waawda

+++ Discord/base/views.py
@@ -3,13 +3,17 @@ from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 # Create your views here.
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method  == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -24,13 +28,28 @@ def loginPage(request):
             return redirect('home')
         else:
             messages.error(request, 'Username or password does not exist ')
-    context = {}
+    context = {'page':page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
 
+def registerPage(request):
+    form = UserCreationForm
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit = False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registration')
+        
+    return render(request, 'base/login_register.html', {'form':form})
 
 def home (request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -41,13 +60,27 @@ def home (request):
                                 )
     topics = Topic.objects.all()
     room_count = rooms.count() #Wbudowana metoda
+    room_messages = Message.objects.all()
 
-    context  = {'rooms': rooms, 'topics':topics, 'room_count':room_count}
+
+    context  = {'rooms': rooms, 'topics':topics,
+                 'room_count':room_count, 'room_messages': room_messages}
     return render(request, 'base/home.html', context)
 
 def room (request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    room_messages = room.message_set.all() #od najnowszych
+    participants = room.participants.all()
+
+    if request.method == "POST":
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+    context = {'room': room, 'room_messages': room_messages, 'participants': participants}
     return render(request, 'base/room.html', context)
 
 @login_required(login_url='login')
@@ -81,8 +114,23 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj':room})
     
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here!')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj': message})
